requirements.txt
flake8
aws-cdk.core
aws-cdk.aws_lambda
aws-cdk.aws_apigateway
aws-cdk.aws_cognito